new commit on 10 december 2022

+++ src/store/actions/index.js
@@ -6,6 +6,7 @@ import {
   LIKE_POST,
   UPDATE,
 } from "../constants/actionTypes";
+import cogoToast from "cogo-toast";
 
 // Action Creators
 
@@ -13,7 +14,11 @@ export const getPosts = () => async (dispatch) => {
   try {
     const { data } = await api.fetchPosts();
 
-    dispatch({ type: FETCH_ALL, payload: data.data });
+    if (data?.success) {
+      dispatch({ type: FETCH_ALL, payload: data.data });
+    } else {
+      cogoToast.error(data?.message);
+    }
   } catch (error) {
     console.log(error.message);
   }
@@ -23,7 +28,12 @@ export const createPost = (post) => async (dispatch) => {
   try {
     const { data } = await api.createPost(post);
 
-    dispatch({ type: CREATE, payload: data.data });
+    if (data?.success) {
+      dispatch({ type: CREATE, payload: data.data });
+      cogoToast.success(data?.message);
+    } else {
+      cogoToast.error(data?.message);
+    }
   } catch (error) {
     console.log(error.message);
   }
@@ -33,7 +43,12 @@ export const updatePost = (id, post) => async (dispatch) => {
   try {
     const { data } = await api.updatePost(id, post);
 
-    dispatch({ type: UPDATE, payload: data.data });
+    if (data?.success) {
+      dispatch({ type: UPDATE, payload: data.data });
+      cogoToast.success(data?.message);
+    } else {
+      cogoToast.error(data?.message);
+    }
   } catch (error) {
     console.log(error);
   }
@@ -41,9 +56,14 @@ export const updatePost = (id, post) => async (dispatch) => {
 
 export const deletePost = (id) => async (dispatch) => {
   try {
-    await api.deletePost(id);
+    const { data } = await api.deletePost(id);
 
-    dispatch({ type: DELETE, payload: id });
+    if (data?.success) {
+      dispatch({ type: DELETE, payload: id });
+      cogoToast.success(data?.message);
+    } else {
+      cogoToast.error(data?.message);
+    }
   } catch (error) {
     console.log(error);
   }
@@ -53,7 +73,12 @@ export const likePost = (id) => async (dispatch) => {
   try {
     const { data } = await api.likePost(id);
 
-    dispatch({ type: LIKE_POST, payload: data.data });
+    if (data?.success) {
+      dispatch({ type: LIKE_POST, payload: data.data });
+      cogoToast.success(data?.message);
+    } else {
+      cogoToast.error(data?.message);
+    }
   } catch (error) {
     console.log(error);
   }
